requirements/base.txt
Django==4.2.3
python-decouple==3.8

psycopg2==2.9.5
djangorestframework==3.14.0

django-filter==23.2
django-extensions==3.2.1
django-cors-headers==4.1.0
django-storages==1.13.1
django-jazzmin==2.6.0
django-jazzmin-admin-rangefilter==1.0.0
djangorestframework-simplejwt==5.2.2
django-debug-toolbar==4.2.0

coreapi==2.3.3

Pillow==10.0.0
requests==2.31.0

boto3==1.26.0
attrs==22.1.0

gunicorn==20.1.0

pyotp==2.8.0
qrcode==7.4.2
PyJWT==2.8.0
python-jose==3.3.0